Ask user to select required permission for authorisation
This change adds a user prompt to select 'read', 'write', or
'delete' when they're retrieving the Access Token.

It also switches to using the SymphonyStyle class for console
input and output, as this gives us a few easy to use methods
and so reduces complexity in this class.

The config file path is also changed to be an absolute path,
both for the default and when it's supplied with the --config
CLI option, because there were errors when running from a
location outside of the base directory.

Refs: #7