New CLI option for 'move-after-upload'
This adds the --move (or -m) option for providing a destination
directory for uploaded files to be moved to after upload. It
removes this option from the config file (and so is a non-BC
change).

If --move is not specified, the current default behaviour of not
moving files is retained.