Set up continuous-integration fro coding standards checking
This adds a Travis config file so that we can turn on CI for
coding standards (because I keep mucking things up ;) ).

It also includes the composer.lock file as this makes repeatable
builds possible (it means that each `composer install` will do the
same thing, and that only committed updates will result in version
increases).